backend/requirements.txt
fastapi
uvicorn
pymongo
python-dotenv
passlib[bcrypt]
requests
pydantic[email]
pip freeze > requirements.txt